+++ reporter/parsers.py
@@ -2,25 +2,21 @@ class JiraParser:
     """A class responsible for parsing a JIRA API response."""
 
     @staticmethod
-    def get_issues_with_state(state: str, issues_list: dict) -> list:
+    def filter_out_important_data(issues_list: dict) -> list:
         """
-        Return issues that have a specific state.
+        Flatter the dictionary to only important information.
 
-        :param state: a state name
         :param issues_list: issues from the API
         """
-        issues = []
-        for issue in issues_list['issues']:
-            status = issue['fields']['status']['name']
-            if status == state:
-                info = {
-                    'id': issue['id'],
-                    'key': issue['key'],
-                    'title': issue['fields']['summary'],
-                    'status': status,
-                    'self': issue['self'],
-                }
-                issues.append(info)
+        issues = [
+            {
+                'id': issue['id'],
+                'key': issue['key'],
+                'title': issue['fields']['summary'],
+                'status': issue['fields']['status']['name'],
+                'self': issue['self'],
+            } for issue in issues_list['issues']
+        ]
         return issues
 
     def parse_pull_request_info(self, tickets_info_response: list) -> list:
@@ -39,6 +35,7 @@ class JiraParser:
                 )
                 if assigned_pull_requests:
                     info = {
+                        'key': detail['key'],
                         'title': detail['title'],
                         'pull_requests': assigned_pull_requests,
                     }
@@ -50,6 +47,8 @@ class JiraParser:
         """
         Return a list of pull requests with a given status.
 
+        Only pull requests which have assigned reviewers that did not give an approval are returned.
+
         :param status: status of pull requests that should be returned
         :param pull_requests: list of dictionaries that contain information
             about pull requests assigned to a ticket
@@ -57,10 +56,12 @@ class JiraParser:
         opened_pull_requests = []
         for pull_request in pull_requests:
             if pull_request['status'] == status:
-                info = {
-                    'author': pull_request['author']['name'],
-                    'url': pull_request['url'],
-                    'reviewers': pull_request['reviewers'],
-                }
-                opened_pull_requests.append(info)
+                reviewers = list(filter(lambda x: not x['approved'], pull_request['reviewers']))
+                if reviewers:
+                    info = {
+                        'author': pull_request['author']['name'],
+                        'url': pull_request['url'],
+                        'reviewers': [x['name'] for x in reviewers],
+                    }
+                    opened_pull_requests.append(info)
         return opened_pull_requests